pyproject.toml
[tool.poetry]
name = "aws-bedrock"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
ruff = "^0.5.0"
boto3 = "^1.34.138"
streamlit = "^1.36.0"
langchain = "^0.2.6"
langchain-experimental = "^0.0.62"
langchain-aws = "^0.1.9"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"
transformers = "^4.34.0"
pypdf2 = "^3.0.1"
tensorflow = "^2.16.2"
tf-keras = "^2.16.0"
google-search-results = "^2.4.2"
tavily-python = "^0.3.3"
psycopg2-binary = "^2.9.9"
psycopg = "^3.2.1"
psycopg-binary = "^3.2.1"
pgvector = "^0.3.0"
anthropic = "^0.30.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
